Fixed caching sha2 password bug where password is xored with scramble would cause an invalid index error due to password having greater length than scramble.
