+++ lib/auth_41.js
@@ -57,17 +57,15 @@ function xor(a, b) {
   return result;
 }
 
+exports.xor = xor;
+
 function token(password, scramble1, scramble2) {
   // TODO: use buffers (not sure why strings here)
   if (!password) {
     return Buffer.alloc(0);
   }
   const stage1 = sha1(password);
-  return exports.calculateTokenFromPasswordSha(
-    stage1,
-    scramble1,
-    scramble2
-  );
+  return exports.calculateTokenFromPasswordSha(stage1, scramble1, scramble2);
 }
 
 exports.calculateTokenFromPasswordSha = function(
@@ -82,12 +80,7 @@ exports.calculateTokenFromPasswordSha = function(
 
 exports.calculateToken = token;
 
-exports.verifyToken = function(
-  publicSeed1,
-  publicSeed2,
-  token,
-  doubleSha
-) {
+exports.verifyToken = function(publicSeed1, publicSeed2, token, doubleSha) {
   const hashStage1 = xor(token, sha1(publicSeed1, publicSeed2, doubleSha));
   const candidateHash2 = sha1(hashStage1);
   return candidateHash2.compare(doubleSha) === 0;
@@ -96,3 +89,23 @@ exports.verifyToken = function(
 exports.doubleSha1 = function(password) {
   return sha1(sha1(password));
 };
+
+function xorRotating(a, seed) {
+  console.log(' OH MA GAWD');
+  if (!Buffer.isBuffer(a)) {
+    a = Buffer.from(a, 'binary');
+  }
+
+  if (!Buffer.isBuffer(seed)) {
+    seed = Buffer.from(seed, 'binary');
+  }
+
+  const result = Buffer.allocUnsafe(a.length);
+  const seedLen = seed.length;
+
+  for (let i = 0; i < a.length; i++) {
+    result[i] = a[i] ^ seed[i % seedLen];
+  }
+  return result;
+}
+exports.xorRotating = xorRotating;
