+++ lib/auth_plugins/caching_sha2_password.js
@@ -4,6 +4,7 @@
 
 const PLUGIN_NAME = 'caching_sha2_password';
 const crypto = require('crypto');
+const { xor, xorRotating } = require('../auth_41');
 
 const REQUEST_SERVER_KEY_PACKET = Buffer.from([2]);
 const FAST_AUTH_SUCCESS_PACKET = Buffer.from([3]);
@@ -20,23 +21,6 @@ function sha256(msg) {
   return hash.digest('binary');
 }
 
-function xor(a, b) {
-  if (!Buffer.isBuffer(a)) {
-    a = Buffer.from(a, 'binary');
-  }
-
-  if (!Buffer.isBuffer(b)) {
-    b = Buffer.from(b, 'binary');
-  }
-
-  const result = Buffer.allocUnsafe(a.length);
-
-  for (let i = 0; i < a.length; i++) {
-    result[i] = a[i] ^ b[i];
-  }
-  return result;
-}
-
 function calculateToken(password, scramble) {
   if (!password) {
     return Buffer.alloc(0);
@@ -48,7 +32,7 @@ function calculateToken(password, scramble) {
 }
 
 function encrypt(password, scramble, key) {
-  const stage1 = xor(
+  const stage1 = xorRotating(
     Buffer.from(`${password}\0`, 'utf8').toString('binary'),
     scramble.toString('binary')
   );
